split get duration job
